Add Task43

+++ Task13/Program.cs
@@ -1,6 +1,6 @@
 ﻿Console.Clear();
 Console.WriteLine("Введите любое число:");
-int a = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToInt32(Console.ReadLine());
 string m = Convert.ToString(a);
 if (m.Length > 2)
